refactoring the role methods and ui
