+++ src/app/app.module.ts
@@ -26,7 +26,6 @@ import { NavbarComponent } from './core/navbar/navbar.component';
 import { MyInboxComponent } from './private/pages/myInbox/my-inbox.component';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { MatOptionModule } from '@angular/material/core';
-import { ProfileComponent } from './private/pages/myProfile/profile.component';
 import { UsersService } from './core/services/user.service';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { SignupComponent } from './public/Pages/signup/signup.component';
@@ -41,22 +40,17 @@ import { MyAccountInfoComponent } from './private/pages/myOverview/my-account-in
 import { FooterComponent } from './shared/footer/footer.component';
 import { MatSelectModule } from '@angular/material/select';
 import { HotToastModule } from '@ngneat/hot-toast';
-import { PublicProfileComponent } from './private/pages/modifyMyPage/public-profile.component';
-import { MyPageViewComponent } from './private/pages/viewMyPage/my-page-view.component';
-import { ModifyMechPageComponent } from './private/pages/modifyMyMechPage/modify-mech-page.component';
+import { ModifyMechPageComponent } from './private/pages/applyPage/modify-mech-page.component';
 @NgModule({
   declarations: [
     AppComponent,
     NavbarComponent,
     routingComponents,
     MyInboxComponent,
-    ProfileComponent,
     SignupComponent,
     DateDisplayPipe,
     MyAccountInfoComponent,
     FooterComponent,
-    PublicProfileComponent,
-    MyPageViewComponent,
     ModifyMechPageComponent,
   ],
