+++ src/app/private/pages/applyPage/modify-mech-page.component.css
@@ -1,4 +1,8 @@
-.profile-image > img {
+:host {
+    padding: 32px;
+  }
+  
+  .profile-image > img {
     border-radius: 100%;
     object-fit: cover;
     object-position: center;
@@ -14,4 +18,13 @@
     position: absolute;
     bottom: 10px;
     right: 0;
+  }
+  
+  .form-container {
+    max-width: 450px;
+  }
+  
+  .row {
+    display: flex;
+    gap: 16px;
   }
