+++ src/app/private/pages/applyPage/modify-mech-page.component.html
@@ -1,0 +1,19 @@
+<div class="form-container mat-elevation-z5 text-center" *ngIf="user$ | async as user">
+    <h2>Apply to become a contractor</h2>
+    <img src="https://txt.1001fonts.net/img/txt/dHRmLjcyLjAwMDAwMC5SMjhnVFdWamFHRnVhV00sLjE,/raleway.extrabold.webp" width="150" height="50"> 
+    <br><br>
+    <h4>start repairing cars now!</h4>
+    <form class="margin-top" [formGroup]="applyRoleForm">
+      <mat-form-field>
+        <input type="text" matInput formControlName="description"
+          placeholder="Description (tell us about yourself bud)" />
+      </mat-form-field>
+      <div class="row">
+        <mat-form-field>
+          <input type="zipcode" matInput formControlName="zip" placeholder="Zip code" />
+        </mat-form-field>
+      </div>
+      <!--Im testing this to be create profile button, will be implementing a visibility guard in future, refactoring comes at a complexity cost.-->
+      <button mat-stroked-button color="primary" (click)="applyRole()" class="margin-top"> send</button>
+    </form>
+  </div>
