+++ src/app/private/pages/applyPage/modify-mech-page.component.ts
@@ -1,12 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
 import { HotToastService } from '@ngneat/hot-toast';
-import { User } from 'firebase/auth';
-import { concatMap, from, Observable, of, switchMap } from 'rxjs';
 import { AuthenticationService } from 'src/app/core/auth/auth.service';
-import { ImageUploadService } from 'src/app/core/services/image-upload.service';
 import { UsersService } from 'src/app/core/services/user.service';
-
 import { PostsService } from 'src/app/core/services/posts.service';
 
 @Component({
@@ -19,7 +15,7 @@ export class ModifyMechPageComponent implements OnInit {
   user$ = this.postsService.currentUserProfile$;
   userAuth$ = this.authService.currentUser$;
 
-  CommunityProfileForm = new FormGroup({
+  applyRoleForm = new FormGroup({
     uid: new FormControl('', {
       nonNullable: true,
     }),
@@ -46,7 +42,6 @@ export class ModifyMechPageComponent implements OnInit {
 
   constructor(
     private authService : AuthenticationService, 
-    private imageUploadService: ImageUploadService, 
     private toast: HotToastService,
     private usersService: UsersService,
     private postsService: PostsService,
@@ -57,29 +52,15 @@ export class ModifyMechPageComponent implements OnInit {
     this.postsService.currentUserProfile$
       //.pipe(untilDestroyed(this)
       .subscribe((user) => {
-        this.CommunityProfileForm.patchValue({ ...user });
+        this.applyRoleForm.patchValue({ ...user });
       });
     }
 
-    uploadImage(event: any, user: User) {
-      this.imageUploadService.uploadImage(event.target.files[0], `images/profile/${user.uid}`).pipe(
-        this.toast.observe({
-          loading: 'Uploading Proflile Image...',
-          success: 'Profile Image Uploaded Successfully',
-          error: 'There was an error uploading',
-        }),
-        concatMap((photoURL) => this.authService.updateProfileData(({ photoURL })))
-        ).subscribe();
-      }
-
-  savePublicProfile() {
-    const {uid, ...data} = this.CommunityProfileForm.value;
-
-    if (!uid) {
-      return; /* Error message portion if UID is undefined for whatever reason. - Anthony */
-    }
+  applyRole() {
+    const {uid, ...data} = this.applyRoleForm.value;
+    if (!uid) { return; }
     //This will add the mechanic role and create their public mechanic profile
-    this.postsService.createPublicProfile({
+    this.usersService.updateUser({
       uid, ...data,
       role: 'mechanic'
     }).pipe(this.toast.observe({
