--- src/app/private/pages/modifyMyMechPage/modify-mech-page.component.css
@@ -1,18 +1,0 @@
-/*These css files are to format the profile picture, please do not forge to import to new components that use profile pic... */
-.profile-image>img {
-    border-radius: 100%;
-    object-fit: cover;
-    object-position: center;
-}
-
-.profile-image {
-    position: relative;
-    width: 150px;
-    margin: auto;
-}
-
-.profile-image>button {
-    position: absolute;
-    bottom: 10px;
-    right: 0;
-}
