+++ src/app/private/pages/myInbox/my-inbox.component.html
@@ -36,7 +36,7 @@
                     <p mat-line>{{ chat.lastMessage }}</p>
                     <mat-divider></mat-divider>
                 </mat-list-option>
-            </mat-selection-list>|
+            </mat-selection-list>
         </div>
     </div>
