+++ src/app/private/pages/myOverview/my-account-info.component.html
@@ -29,7 +29,8 @@
     </p>
 
     <ul class="menu-list">
-      <li><a class="menu-item" (click)="openSideNav($event, 'Payments')">Payments</a></li>
+      <li><a class="menu-item" (click)="openSideNav($event, 'Payments')">Payments</a></li> <br>
+      <h5><a class="menu-item" routerLink="/applyRole">Become a mechanic</a></h5>
     </ul>
   </mat-drawer>
   <!--End of side-navbar menu -Jim-->
@@ -40,8 +41,34 @@
 
     <!--Top of Dashboard Section -->
     <div id="Dashboard" class="menu-content">
-      Dashboard.
+      <div class *ngIf="userProf$ | async as user"> <!-- this communicates to db via 'Community profiles' -->
+        <div class *ngIf="user$ | async as userAuth"> <!--this communicates to db via 'users' - [jsb]-->
+            <div class="container">
+                <div class="row">
+                    <div class="col-12 col-sm-6 col-md-4 col-lg-3">
+                        <div class="our-team">
+                            <div class="picture"> <img width="120" height="120" [src]="userAuth.photoURL ?? 'assets/images/placeholderProfile.png'" />
+                            </div>
+                                <div class="team-content">
+                                    <h1 class="name"> <span class="user-name"> {{ user.displayName}} </span>
+                                    </h1>     
+                                        <h3 class="name">
+                                            <span>About '{{user.firstName}}', {{user.description}} 
+                                            </span>
+                                        </h3>
+                                            <h5 class="name"> A Go-Mechanic 
+                                                <span class="user-info"> {{ user.role}}, near </span> 
+                                                <span class="user-info"> {{ user.zip }} </span>
+                                            </h5>
+                                </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
     </div>
+    
+      </div>
     <!--End of Dashboard Section-->
 
     <!--Top of Account Management Section -->
@@ -120,17 +147,22 @@
                 placeholder="zip code"
               />
             </mat-form-field>
+
+            <mat-form-field>
+              <input type
+                type="text"
+                matInput
+                formControlName="description"
+                placeholder="description"
+              />
+            </mat-form-field>
+
+
           </div>
 
           <div>
             <button type="submit" style ="margin-top: 2px" mat-raised-button color="primary">
               Save
-            </button>
-          </div>
-
-          <div *ngIf="userProf$ | async as userAuth" >
-            <button type="submit" (ngSubmit)="saveCommunityProfile(userAuth)" mat-raised-button color="primary" style="margin-top: 5px">
-              Update Community Profile
             </button>
           </div>
         </form>
@@ -194,6 +226,15 @@
       Payment Section <!--Placeholder for Payment Section -Jim-->
     </div>
     <!--End of Payment Section-->
+
+
+
+
+
+
+
+
+
   </mat-drawer-content>
   <!--End of side-nav content holders-->
