+++ src/app/private/pages/myOverview/my-account-info.component.ts
@@ -57,6 +57,9 @@ export class MyAccountInfoComponent implements OnInit {
     address: new FormControl('', {
       nonNullable: true,
     }),
+    description: new FormControl('', {
+      nonNullable: true,
+    }),
     zip: new FormControl('', {
       nonNullable: true,
     }),
@@ -118,38 +121,29 @@ export class MyAccountInfoComponent implements OnInit {
       ).subscribe();
     }
 
-    //Saves to 'users' collection firestore
-    saveProfile(){  
+  //Saves to 'users' collection firestore and verifies role...needs refractoring
+  saveProfile(){  
       const {uid, ...data} = this.profileForm.value;
       
       //This checks if a user already has a 'mechanic role', if not it will continue with saveCommunityProfile method.
       if(this.profileForm.value.role === 'mechanic'){
-        if (!uid) {
-          return; /* Error message portion if UID is undefined for whatever reason. - Anthony */
-        }
-        this.usersService.updateUser({
-          uid, ...data,
-          role: 'mechanic'
-        })
+        if (!uid) { return;}
+        this.usersService.updateUser({uid, ...data, role: 'mechanic' })
         .pipe(this.toast.observe({
-          loading: 'Updating data...',
-          success: 'Your secured mechanic account been updated',
+          success: 'Mechanic profiles have been updated',
           error: 'There was an error in updating the data.'
         })
         )
         .subscribe();
       }
       else{
-        if (!uid) {
-        return; /* Error message portion if UID is undefined for whatever reason. - Anthony */
-        }
+        if (!uid) { return; }
       this.usersService.updateUser({
         uid, ...data,
         role: 'user'
       })
       .pipe(this.toast.observe({
-        loading: 'Updating data...',
-        success: 'Data has been successfully updated!',
+        success: 'Profiles have been successfully updated',
         error: 'There was an error in updating the data.'
       })
       )
@@ -157,42 +151,6 @@ export class MyAccountInfoComponent implements OnInit {
     }
   }
 
-//Saves to 'CommunityProfile' collection firestore
-  saveCommunityProfile(user: User) {
-    const {uid, ...data} = this.CommunityProfileForm.value;
-  if(this.CommunityProfileForm.value.role === 'mechanic'){
-    if (!uid) {
-      return; /* Error message portion if UID is undefined for whatever reason. - Anthony */
-    }
-    this.postsService.createPublicProfile({
-      uid, ...data,
-      role: 'mechanic'
-    })
-    .pipe(this.toast.observe({
-      loading: 'Updating data...',
-      success: 'Your mechanic profile has updated.',
-      error: 'There was an error in updating the data.'
-    })
-    )
-    .subscribe();
-  }
-  else{
-    if (!uid) {
-    return; /* Error message portion if UID is undefined for whatever reason. - Anthony */
-    }
-  this.postsService.createPublicProfile({
-    uid, ...data,
-    role: 'user'
-  })
-  .pipe(this.toast.observe({
-    loading: 'Updating data...',
-    success: 'Your profile has been updated!',
-    error: 'There was an error in updating the data.'
-  })
-  )
-  .subscribe();
-}
-}
 
 //Code for the new dropdown menu to select account type - Anthony
   // Function I created in order to allow me to change tabs within the side-nav bar instead of using page components. This took forever! - Anthony
@@ -209,6 +167,7 @@ export class MyAccountInfoComponent implements OnInit {
       document.getElementById(tabName)!.style.display = "block";
       event.currentTarget.className += " is-active";
   }
+
 
   addCar(){
     //Placeholder for now
