--- src/app/private/pages/myProfile/profile.component.html
@@ -1,77 +1,0 @@
-<div
-  class="form-container mat-elevation-z5 text-center"
-  *ngIf="user$ | async as user"
->
-  <h1>Secured User Profile</h1>
-  <div class="profile-image">
-    <img
-      width="120"
-      height="120"
-      [src]="user.photoURL ?? 'assets/images/placeholderProfile.png'"
-    />
-    <button mat-mini-fab (click)="inputField.click()">
-      <mat-icon>edit</mat-icon>
-    </button>
-  </div>
-  <input #inputField hidden type="file" (change)="uploadImage($event, user)" />
-  <form class="margin-top" [formGroup]="profileForm" (ngSubmit)="saveProfile()">
-    <div class="row">
-      <mat-form-field>
-        <input
-          type="text"
-          matInput
-          formControlName="firstName"
-          placeholder="First Name"
-        />
-      </mat-form-field>
-      <mat-form-field>
-        <input
-          type="text"
-          matInput
-          formControlName="lastName"
-          placeholder="Last Name"
-        />
-      </mat-form-field>
-    </div>
-
-    <div class="row">
-      <mat-form-field>
-        <input
-          type="text"
-          matInput
-          formControlName="displayName"
-          placeholder="Display Name"
-        />
-      </mat-form-field>
-      <mat-form-field>
-        <input
-          type="number"
-          matInput
-          formControlName="phone"
-          placeholder="Phone"
-        />
-      </mat-form-field>
-    </div>
-    <mat-form-field>
-      <input
-        type="text"
-        matInput
-        formControlName="address"
-        placeholder="Address"
-      />
-    </mat-form-field>
-
-    <mat-form-field>
-      <input type
-        type="text"
-        matInput
-        formControlName="zip"
-        placeholder="zip code"
-      />
-    </mat-form-field>
-
-    <button type="submit" class="margin-top" mat-raised-button color="primary">
-      Save
-    </button>
-  </form>
-</div>
