+++ src/app/private/private.module.ts
@@ -3,13 +3,10 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 //Routing Imports
-import { PrivateRoutingModule } from './private-routing.module';
-import { PrivateComponent } from './components/private.component';
 import { SharedModule } from '../shared/shared.module';
 
 //Page Imports
 import { RouterModule } from '@angular/router';
-import { MyAccountInfoComponent } from './pages/myOverview/my-account-info.component';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDividerModule } from '@angular/material/divider';
